Fix page navigation via back/forward buttons
This commit fixes an issue with page naviation by removing Turbolinks and
some other extraneous js stuff. When navigating between pages with the
back and foward button, the page would appear to go blank.  However, the
elm debugger continued to work and I could still see the API was occuring
in the logs.  Turbolinks was getting in the way of elm app handling
browser navigation events. Turbolinks was installed by default when I
started this app as a plain rails app.