Load current user from cookie and scope actions based on current user
On the elm side, the application now tries to load a user from the
back-end based on the remember_token cookie present. If the cookie is
not present, then the app is initialized in the signed out state.

On the rails side, a new `Api::BaseController` was introduced that all
other API controllers inherit from. This controller provides a
before_action that requires a logged in user for all actions by default.

A new `Api::CurrentUsers` controller was added to provide the front-end
with the current user.

The `Api::PlantsController` and `Api::WateringsController` now scope
queries to the current user.

Finally, the requests specs were fleshed out and cleaned up with
some helper functions and factories.