Show last watering time as "# days ago" string
This commit changes the way the last watering date for each plant is
displayed in the plant list. Previously the date with DD/MM/YYYY format
was shown. Now the back-end passes last watering date as Posix value to
front end where conversions and formatting are done.

This commit also introduces a new DateAndTime module to the elm app.