Create plant care frequency
This commit adds the ability to specify how often you want to check a
given plant.

This will allow us to add some inputs to the front end so we can specify
care frequency on plant creation (e.g., 1 week, 3 days).

I also refactored the waterings/checks into a single class
`PlantCareEvent`. They seemed to overlap in functionality/data. Using
the new model, we can now calculate when the plant should be checked
next based the last time a plant was cared for (watered or checked)