Don't show sign in page when user is signed in
This commit changes the way user loading happens in the main module.
Instead of always initially showing the sign in page when the current
user check hasn't yet finished, a new three state `Loadable` type was
introduced. The `currentUser` is initially set to `Loading` and the
corresponding `LoadingPage` is shown until the `currentUser` is resolved
to `Success` or `None`. The user is then redirected to the appropriate
page.