Change plant list styling
This commit updates the plant list styling and functionality. A new
design was introduced that is a little more lighter-weight than the
previous card-based implementation. The plant list is now ordered by
`next_check_time`, with plants that are closest to the check-in or
overdue for a check-in are at the top. Additionally, plants that are
overdue for a check-in have a little clock icon next to them.