Fix stuck loading state for user
This commit fixes an issue where the loading state for a user would not
successfully resolve and render the appropriate page. When no user is
logged in, navigating to the "/sign_in" url directly would result in a
situation where the proper page would not be loaded. This commit changes
the behavior around the `RecievedCurrentUserResponse` message so that
the proper page is loaded when a non-successful code is returned by the
server.