Don't cycle CSRF token on sign in
This commit changes the configuration of clearance so that it doesn't
cycle the csrf token on sign in. Because of the way we're serving up the
front-end, the csrf token is only loaded once at app initialization.
This created a situation where the csrf token passed into the app would
be invalidated straight away.