Add edit plant page
This commit adds the ability to edit a plant from the plant details
page. Users can click the previously added "edit" button to visit the
plant form. The plant form was refactored to support both create and
update actions.

This commit also sneaks in a feature test for deleting a plant