Fix undefined croppieTarget element when taking photo
This commit fixes a bug when trying to add an avatar to a plant with a
photo taken straight from the android camera while using android chrome.
Upon returning to the page after taking a picture, the croppieTarget
element would not be found and Croppie would not initialize. This commit
adds some code to recursively check for the element until it appears,
and then initialize Croppie.