Add rotate buttons to cropper modal, read exif data
This commit adds rotate buttons to the cropper modal. The cropper also
now reads the photos exif data (if there is any) and set the proper
orientation at init. This is useful for photos from a mobile device when
the user if taking images with their device at different orientations.
The user wont have to manually rotate the image.