Make reminder mailer a little nicer
This commit adds some basic styling to the reminder mailer. The
`premailer` gem was added to avoid inline styles in each mailer.

This commit also fixes a bug in the `needs_care` scope used in the
`ReminderJob` to grab plants that need care. It now properly returns a
unique list of plants that need attention.