Add plant gardens
This commit sets up a new `Garden` model and rearranges some
relationships. Users now belong to a garden, and are associated with
plants through their garden.

Plants now have a `garden_id` instead of a `user_id`, and the user
who added the plant to a garden is recorded as `added_by_id`.

Check ins are now associated directly with the user who created it
using the 'performed_by_id` column.

Currently, users can only belong to a single garden at once, and it's
automatically created for them on sign up. A future commit will add some
features around belonging to multiple gardens and managing them.