Create a many-to-many relationship between users and gardens
Gardens now relate to users with a HABTM relationship, and have an
owner. Users can own multiple gardens and be members of multiple
gardens.

The API endpoints were updated so that plants are created and fetched
according to their garden where necessary. The resources are shallowly
nested, so only index and create actions for plants are nested under
garden resources. Accordingly, the front-end was updated to point to the
correct URLs using the necessary resource IDs.

None of the features around creating and managing gardens have been
exposed to the front end yet. For now, users are created with a single
owned garden which is assigned to them as their default garden for use
by the front end.