Add garden select to front end
With this commit, users can now select gardens they belong to and see
and create plants in those gardens. There's a new slide-out menu from
which users can see and select from the gardens they own and belong to.

The front-end routes were updated to follow the shallowly nested rails
routes. Plants are now created in the context of the currently selected
garden.

The next two pieces that'll be put in place for this are CRUD for
gardens and the ability to invite other users to gardens.