Extract menu into own module
In order to separate the logic of managing menu state from the logic in
the main module, this commit pulls the menu into its own module. The
menu now has its own Msg type, and init, update and view functions. In
the main module, the model has a new field `menu` of type Menu.