Add new gardens from menu
This commit adds the ability for users to create new gardens from the
menu. A new + button opens the form, which includes a basic validation
for the presence of a name. Users are redirected to the garden upon
successful creation.

I also extracted some of the menu styling out into its own file.