Open up API to mobile client
This commit makes a number of changes to the API to allow access to both
mobile and browser clients.

A separate auth token, `mobile_api_token`, was created to use
exclusively for mobile in order to keep the sessions separate (i.e.,
signing out on mobile wont log out web and vice versa).

A `Session-Type` header is used to distinguish between the two client
types/authentication schemes.