Fix .need_care query on Plant model
This commit fixes an issue with the `.need_care` query that caused it to
include plants that did not need care. The issue was caused by the query
looking at all past check ins for a plant and comparing their creation
times to the next check time, even if a plant might have had a check in
that satisfied the check frequency. The plant's past check ins would not
meet the criteria and cause it to be included.