Soft delete plants
This commit replaces the previous behavior of hard deleting plants. When
a user deletes a plant, a `deleted_at` flag is now set. The plant no
long shows up in the plant list and the user will be redirected to the
plant list if they try to manually navigate to the plant's detail page.

Follow up work will add a "graveyard" for dead plants.