Notices (#18)
This commit introduces functionality to give users feedback on their actions using a flash message.

Two new types are introduced: Notice and NoticeQueue

The Notice type represents the message and kind of message (alert, warning, etc.) being presented to the user

The NoticeQueue type represents a list of Notice in case many messages need to be presented to the user. Each time a Notice is added to the queue, a timeout is started. After the timeout, the message on the front of the queue is removed so the next Notice in queue, if any, can be shown.